Improve short-sha to better match git outputs
This is a naive implementation that simply counts commits for given SHA.

Git actually counts all the objects in repository, so a better
approximation would at least use the repo's default branch instead of
current SHA, but I don't expect this action to be used on old commits
much.

There's still no checking if provided short-sha is unique or not.